apps/stock-dashboard/backend/requirements.txt
# backend/requirements.txt
Flask==3.1.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
yfinance==0.2.66
pandas==2.2.0
requests==2.31.0
numpy==1.26.3
werkzeug==3.1.0

